Former-commit-id: 39e32142065b430afb595300f2ad3f56b62fc502

+++ unisim_lib/unisim/component/cxx/processor/hcs12x/types.hh
@@ -48,7 +48,7 @@ typedef uint32_t physical_address_t;
 typedef uint16_t reg_t;
 typedef uint64_t service_address_t;
 typedef uint8_t	 page_t;
-typedef double clock_t;
+typedef uint64_t clock_t;
 
 class TOWNER {
 public:
